temp

+++ particle_filter/agents/src/measurements.rs
@@ -1,5 +1,5 @@
 use crate::swarm_element::SwarmElement;
 
 pub trait Measurements {
-    fn ranging(&self, swarm_element: &SwarmElement, std_raning: f32) -> f32;
+    fn ranging(&self, swarm_element: &SwarmElement, std_raning: f64) -> f64;
 }
